Amend composability
IMO all of the frameworks are much closer by this metric.
I think for BB+MN it's probably true that it's *easier* to
do something wrong/messy since it's overall the least
opinionated, but that doesn't prevent it from being adequate
to make a highly composable application.

Another factor that's harder to quantify is institutional
knowledge -- as a team we've got a lot of hard-earned
know-how for building BB apps, and MN complements that style
very nicely.